move qualifications API into v1 namespace
