+++ config/routes.rb
@@ -210,7 +210,9 @@ Rails.application.routes.draw do
     end
 
     namespace :teacher_record_service, path: "teacher-record-service", defaults: { format: :json } do
-      resources :qualifications, path: "npq-qualifications", only: %i[show], param: :trn
+      namespace :v1 do
+        resources :qualifications, path: "npq-qualifications", only: %i[show], param: :trn
+      end
     end
   end
